PQ distortion-based selection, proximity metric, code_size fix, analyze command
Adaptive PQ selection:
- Replace recall-floor with distortion (median |1 - pq_dist/true_dist|) as
  the PQ quality signal. Distortion is geometrically grounded and robust to
  near-duplicate clustering, unlike row-id recall which is capped by tie
  structure on clustered datasets.
- Selection: min cost (m) among configs with distortion <= max_distortion
  (default 0.05), tie-break by table size. Removes cache_info (sysctl/sysfs
  detection) and residency factors (1.0/1.5/1.9) -- table size alone encodes
  the cache-residency information.
- Probe reports band_recall, ties@10, ties@30 as diagnostics.
- --pq-recall-floor flag replaced by --pq-max-distortion.

Proximity metric in benchmark:
- For each returned result, computes ratio = d_target / d_result against the
  GT k-th NN distance. Reports in-band fraction, mean_ratio, and miss_ratio
  percentiles. On clustered data, proximity is the meaningful quality signal
  (recall@10 is capped by ties); on non-clustered data both agree.

Critical bug fix: code_size uint8_t overflow
- PQ code_size (m * bits / 8) was stored as uint8_t throughout the storage
  layer. For m=256/8-bit, code_size=256 wraps to 0, causing zero-stride
  pointer arithmetic and segfault in lut_distance. Widened to uint32_t in
  VamanaCore, FlatNodeStore, PagedNodeStore, MemGraph, and all call sites.

SIFTsmall test fixtures:
- Committed real SIFTsmall (10K vectors, 128d) as .fbin/.gt with real
  distances. Replaced fragile synthetic-noise recall tests with stable
  SIFTsmall-based assertions (recall ~0.95, threshold 0.80).
- Dropped recall assertion in test_memgraph (SIFTsmall too small for MemGraph
  delegation testing -- caches 98% of nodes). Architectural checks retained.
- fvecs_to_fbin converter: added --base/--queries to compute real L2-sq
  distances in GT (was zero-filled).

sextant analyze command:
- Standalone pre-build PQ sensitivity analysis. Builds mini Vamana graphs on
  the 20K probe pool at each eligible (m, bits) config, searches with rerank,
  reports recall@10 per config. Reveals whether the dataset is PQ-sensitive
  (SIFT: recall varies with m) or PQ-insensitive (geolocation: recall flat
  across configs due to tie-capping).
- Pool size governed by three constraints: navigability (N>=4K), statistical
  power (>=200 queries), density representativeness (dim>=32). See
  docs/sensitivity_probe.md.

Documentation:
- README: rewritten adaptive PQ section (distortion, cost=m, no residency
  factors), new Evaluation section defining recall@k, new Why not row-id
  recall section explaining ties/band_recall/clustering.
- docs/sensitivity_probe.md: design rationale and edge-case handling.
- All se_base references replaced with generic geolocation embeddings.

+++ src/engine/search.cpp
@@ -241,7 +241,7 @@ void Engine::load_sidecars() {
     // Reconstruct the final node size from the resolved params.
     node_size_ = VamanaCore::static_node_size(
         params.R, params.inline_pq_count,
-        static_cast<uint8_t>(code_size_));
+        code_size_);
 
     // --- Reconstruct the VamanaCore with the loaded params ---
     VamanaParams vparams;
@@ -307,7 +307,7 @@ void Engine::load_sidecars() {
 
         paged_store_ = std::make_unique<PagedNodeStore>(
             index_path_ + ".graph", index_path_ + ".codes",
-            node_size_, static_cast<uint8_t>(code_size_),
+            node_size_, code_size_,
             std::max(1u, std::thread::hardware_concurrency()),
             cache_bytes);
     }
@@ -323,7 +323,7 @@ void Engine::load_sidecars() {
         }
         memgraph_ = std::make_unique<MemGraph>(
             index_path_ + ".graph", index_path_ + ".codes",
-            node_size_, static_cast<uint8_t>(code_size_),
+            node_size_, code_size_,
             static_cast<uint32_t>(count_), eps, /*num_hops=*/3);
         memgraph_->set_backing(paged_store_.get());
